Clarify ClusterResources note
The note `clusterResources is always included` seems to imply that you don't need to add anything, but you actually need to collect the things from a namespace for this to work.

I would guess this note about `clusterResources is always included` is in other places that could benefit from the same clarification (e.g. I hit the same issue for the DeploymentStatus collector).